Moved into new flags directory under feature
