+++ packages/ui/src/features/CourtCaseDetails/flags/canUseCourtDateReceivedDateMismatchFilters.ts
@@ -1,5 +1,5 @@
 import { FORCES_WITH_COURT_DATE_RECEIVED_DATE_MISMATCH_ENABLED } from "config"
-import type User from "./entities/User"
+import type User from "../../../services/entities/User"
 
 export const canUseCourtDateReceivedDateMismatchFilters = ({ featureFlags, visibleForces }: User): boolean => {
   if (!featureFlags.useCourtDateReceivedDateMismatchFiltersEnabled) {
